ai_scoring/requirements.txt
# TigerTrust AI Scoring Service Dependencies

# Core dependencies
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1

# Google Gemini AI
google-generativeai==0.8.3

# Solana blockchain interaction
solana>=0.36.1
solders>=0.21.0
anchorpy==0.21.0

# HTTP requests
requests==2.32.3

# Async support
aiohttp==3.11.11

# Utilities
python-dateutil==2.9.0

# Development tools (optional)
pytest==8.3.4
pytest-asyncio==0.25.2
black==25.1.0
